requirements.txt
streamlit
networkx
matplotlib
pillow
pandas